requirements.txt
Django==4.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
pillow==11.2.1 